Allow master build without publishing to download.eclipse.org

+++ releng/org.eclipse.m2m.atl.update/publish.sh
@@ -12,14 +12,20 @@
 #*******************************************************************************
 # Local ATL publish script for use by the promoter job. It prepares and invokes remote scripts.
 #
-rm -f downloads.sh publish.zip updates.sh
-curl -s -k ${PUBLISH__DOWNLOADS_SH} > downloads.sh
-curl -s -k ${PUBLISH__UPDATES_SH} > updates.sh
-curl -s -k ${PUBLISH__URL} > publish.zip
-ssh genie.atl@projects-storage.eclipse.org rm -f downloads.sh publish.zip updates.sh
-scp downloads.sh genie.atl@projects-storage.eclipse.org:downloads.sh
-scp publish.zip genie.atl@projects-storage.eclipse.org:publish.zip
-scp updates.sh genie.atl@projects-storage.eclipse.org:updates.sh
-ssh genie.atl@projects-storage.eclipse.org chmod +x downloads.sh updates.sh
-ssh genie.atl@projects-storage.eclipse.org bash -ex downloads.sh -v "'${PUBLISH__VERSION}'" -t "'${PUBLISH__BUILD_T}'" -q "'${PUBLISH__QUALIFIER}'" -a "'${PUBLISH__ALIAS}'"
-ssh genie.atl@projects-storage.eclipse.org bash -ex updates.sh -v "'${PUBLISH__VERSION}'" -t "'${PUBLISH__BUILD_T}'" -q "'${PUBLISH__QUALIFIER}'"
+
+if [[ ${PUBLISH__PROMOTE} == "true" ]]
+then
+
+	rm -f downloads.sh publish.zip updates.sh
+	curl -s -k ${PUBLISH__DOWNLOADS_SH} > downloads.sh
+	curl -s -k ${PUBLISH__UPDATES_SH} > updates.sh
+	curl -s -k ${PUBLISH__URL} > publish.zip
+	ssh genie.atl@projects-storage.eclipse.org rm -f downloads.sh publish.zip updates.sh
+	scp downloads.sh genie.atl@projects-storage.eclipse.org:downloads.sh
+	scp publish.zip genie.atl@projects-storage.eclipse.org:publish.zip
+	scp updates.sh genie.atl@projects-storage.eclipse.org:updates.sh
+	ssh genie.atl@projects-storage.eclipse.org chmod +x downloads.sh updates.sh
+	ssh genie.atl@projects-storage.eclipse.org bash -ex downloads.sh -v "'${PUBLISH__VERSION}'" -t "'${PUBLISH__BUILD_T}'" -q "'${PUBLISH__QUALIFIER}'" -a "'${PUBLISH__ALIAS}'"
+	ssh genie.atl@projects-storage.eclipse.org bash -ex updates.sh -v "'${PUBLISH__VERSION}'" -t "'${PUBLISH__BUILD_T}'" -q "'${PUBLISH__QUALIFIER}'"
+
+fi
